Add files via upload
This is a datapack maker written in batch. Easily creating datapacks in minecraft. This is in beta dont think its done because its not. Oh yeah and the item files are for the recipes but the % thought it was a variable so i couldnt do it any other way sorry. Oh yeah and in this beta recipes are completely doomed so dont try or you get bad things in your datapack plus old doesnt work yet idk i havent tested it but i didnt add any yet.